requirements.txt
# Core dependencies
langgraph>=0.2.0
anthropic>=0.18.0
typer>=0.9.0
rich>=13.0.0

# LaTeX processing
pylatexenc>=2.10

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
